Se elimina archivo Menu roto, se crea Menu principal
