--- UniversidadGrupo67/src/newpackage/vistas/Menu.java
@@ -1,428 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package newpackage.vistas;
-
-/**
- *
- * @author Natasha
- */
-public class Menu extends javax.swing.JFrame {
-
-    /**
-     * Creates new form Menu
-     */
-    public Menu() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jMenuItem1 = new javax.swing.JMenuItem();
-        jMenu3 = new javax.swing.JMenu();
-        jMenuBar2 = new javax.swing.JMenuBar();
-        jMenu4 = new javax.swing.JMenu();
-        jMenu5 = new javax.swing.JMenu();
-        jMenu6 = new javax.swing.JMenu();
-        jMenu7 = new javax.swing.JMenu();
-        jMenu8 = new javax.swing.JMenu();
-        jMenu9 = new javax.swing.JMenu();
-        jMenu10 = new javax.swing.JMenu();
-        jMenuBar3 = new javax.swing.JMenuBar();
-        jMenu11 = new javax.swing.JMenu();
-        jMenu12 = new javax.swing.JMenu();
-        jMenu13 = new javax.swing.JMenu();
-        jMenu14 = new javax.swing.JMenu();
-        jMenu15 = new javax.swing.JMenu();
-        jMenu16 = new javax.swing.JMenu();
-        jMenu17 = new javax.swing.JMenu();
-        jMenu18 = new javax.swing.JMenu();
-        jMenu19 = new javax.swing.JMenu();
-        jMenu20 = new javax.swing.JMenu();
-        jMenu21 = new javax.swing.JMenu();
-        jMenu22 = new javax.swing.JMenu();
-        jMenu23 = new javax.swing.JMenu();
-        jMenu25 = new javax.swing.JMenu();
-        jMenu26 = new javax.swing.JMenu();
-        jMenu27 = new javax.swing.JMenu();
-        jMenu28 = new javax.swing.JMenu();
-        jMenu29 = new javax.swing.JMenu();
-        jMenu30 = new javax.swing.JMenu();
-        jMenu31 = new javax.swing.JMenu();
-        jMenu32 = new javax.swing.JMenu();
-        jMenu33 = new javax.swing.JMenu();
-        jMenu34 = new javax.swing.JMenu();
-        jMenu35 = new javax.swing.JMenu();
-        jMenu36 = new javax.swing.JMenu();
-        jMenu37 = new javax.swing.JMenu();
-        jMenu39 = new javax.swing.JMenu();
-        jMenu40 = new javax.swing.JMenu();
-        jMenu41 = new javax.swing.JMenu();
-        jMenu42 = new javax.swing.JMenu();
-        jMenu43 = new javax.swing.JMenu();
-        jMenu44 = new javax.swing.JMenu();
-        jMenu45 = new javax.swing.JMenu();
-        jMenu46 = new javax.swing.JMenu();
-        jPanelPrincipal = new javax.swing.JDesktopPane();
-        jMenuBar1 = new javax.swing.JMenuBar();
-        jmAlumnos = new javax.swing.JMenu();
-        jMIFormularioAlumno = new javax.swing.JMenuItem();
-        jmMateria = new javax.swing.JMenu();
-        jMIFormularioMateria = new javax.swing.JMenuItem();
-        jMenu24 = new javax.swing.JMenu();
-        jMIInscripciones = new javax.swing.JMenuItem();
-        jMManipulacionDeNotas = new javax.swing.JMenuItem();
-        jMenu38 = new javax.swing.JMenu();
-        jMenuItem2 = new javax.swing.JMenuItem();
-        jMenu47 = new javax.swing.JMenu();
-
-        jMenuItem1.setText("jMenuItem1");
-
-        jMenu3.setText("jMenu3");
-
-        jMenu4.setText("File");
-        jMenuBar2.add(jMenu4);
-
-        jMenu5.setText("Edit");
-        jMenuBar2.add(jMenu5);
-
-        jMenu6.setText("jMenu6");
-
-        jMenu7.setText("jMenu7");
-
-        jMenu8.setText("jMenu8");
-
-        jMenu9.setText("jMenu9");
-
-        jMenu10.setText("jMenu10");
-
-        jMenu11.setText("File");
-        jMenuBar3.add(jMenu11);
-
-        jMenu12.setText("Edit");
-        jMenuBar3.add(jMenu12);
-
-        jMenu13.setText("jMenu13");
-
-        jMenu14.setText("jMenu14");
-
-        jMenu15.setText("jMenu15");
-
-        jMenu16.setText("jMenu16");
-
-        jMenu17.setText("jMenu17");
-
-        jMenu18.setText("jMenu18");
-
-        jMenu19.setText("jMenu19");
-
-        jMenu20.setText("jMenu20");
-
-        jMenu21.setText("jMenu21");
-
-        jMenu22.setText("jMenu22");
-
-        jMenu23.setText("jMenu23");
-
-        jMenu25.setText("jMenu25");
-
-        jMenu26.setText("jMenu26");
-
-        jMenu27.setText("jMenu27");
-
-        jMenu28.setText("jMenu28");
-
-        jMenu29.setText("jMenu29");
-
-        jMenu30.setText("jMenu30");
-
-        jMenu31.setText("jMenu31");
-
-        jMenu32.setText("jMenu32");
-
-        jMenu33.setText("jMenu33");
-
-        jMenu34.setText("jMenu34");
-
-        jMenu35.setText("jMenu35");
-
-        jMenu36.setText("jMenu36");
-
-        jMenu37.setText("jMenu37");
-
-        jMenu39.setText("jMenu39");
-
-        jMenu40.setText("jMenu40");
-
-        jMenu41.setText("jMenu41");
-
-        jMenu42.setText("jMenu42");
-
-        jMenu43.setText("jMenu43");
-
-        jMenu44.setText("jMenu44");
-
-        jMenu45.setText("jMenu45");
-
-        jMenu46.setText("jMenu46");
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setPreferredSize(new java.awt.Dimension(1200, 800));
-
-        jPanelPrincipal.setPreferredSize(new java.awt.Dimension(1800, 1200));
-
-        javax.swing.GroupLayout jPanelPrincipalLayout = new javax.swing.GroupLayout(jPanelPrincipal);
-        jPanelPrincipal.setLayout(jPanelPrincipalLayout);
-        jPanelPrincipalLayout.setHorizontalGroup(
-            jPanelPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 662, Short.MAX_VALUE)
-        );
-        jPanelPrincipalLayout.setVerticalGroup(
-            jPanelPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 358, Short.MAX_VALUE)
-        );
-
-        jmAlumnos.setText("Alumno");
-
-        jMIFormularioAlumno.setText("Formulario Alumno");
-        jMIFormularioAlumno.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jMIFormularioAlumnoActionPerformed(evt);
-            }
-        });
-        jmAlumnos.add(jMIFormularioAlumno);
-
-        jMenuBar1.add(jmAlumnos);
-
-        jmMateria.setText("Materia");
-
-        jMIFormularioMateria.setText("Fomulario Materia");
-        jMIFormularioMateria.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jMIFormularioMateriaActionPerformed(evt);
-            }
-        });
-        jmMateria.add(jMIFormularioMateria);
-
-        jMenuBar1.add(jmMateria);
-
-        jMenu24.setText("Administracion");
-
-        jMIInscripciones.setText("Manejo de Inscripciones");
-        jMIInscripciones.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jMIInscripcionesActionPerformed(evt);
-            }
-        });
-        jMenu24.add(jMIInscripciones);
-
-
-        jMManipulacionDeNotas.setText("Manipulacion De Notas");
-        jMManipulacionDeNotas.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jMManipulacionDeNotasActionPerformed(evt);
-            }
-        });
-        jMenu24.add(jMManipulacionDeNotas);
-
-        jMenuBar1.add(jMenu24);
-
-        jMenu38.setText("consulta");
-
-        jMenuItem2.setText("Alumnos por materia");
-        jMenuItem2.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jMenuItem2ActionPerformed(evt);
-            }
-        });
-        jMenu38.add(jMenuItem2);
-
-        jMenuBar1.add(jMenu38);
-
-        jMenu47.setText("salir");
-        jMenu47.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jMenu47ActionPerformed(evt);
-            }
-        });
-        jMenuBar1.add(jMenu47);
-
-        setJMenuBar(jMenuBar1);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanelPrincipal, javax.swing.GroupLayout.DEFAULT_SIZE, 662, Short.MAX_VALUE)
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addComponent(jPanelPrincipal, javax.swing.GroupLayout.DEFAULT_SIZE, 358, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-
-    private void jMIFormularioAlumnoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMIFormularioAlumnoActionPerformed
-        // TODO add your handling code here:
-        FormularioAlumnos ventana=new FormularioAlumnos();
-        jPanelPrincipal.removeAll();
-        jPanelPrincipal.repaint();
-        jPanelPrincipal.add(ventana);
-        ventana.setVisible(true);
-    }//GEN-LAST:event_jMIFormularioAlumnoActionPerformed
-
-    private void jMIFormularioMateriaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMIFormularioMateriaActionPerformed
-        // TODO add your handling code here:
-        FormularioMateria ventana=new FormularioMateria();
-        jPanelPrincipal.removeAll();
-        jPanelPrincipal.repaint();
-        jPanelPrincipal.add(ventana);
-        ventana.setVisible(true);
-    }//GEN-LAST:event_jMIFormularioMateriaActionPerformed
-
-    private void jMIInscripcionesActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMIInscripcionesActionPerformed
-        // TODO add your handling code here:
-        Inscripciones ventana=new Inscripciones();
-        jPanelPrincipal.removeAll();
-        jPanelPrincipal.repaint();
-        jPanelPrincipal.add(ventana);
-        ventana.setVisible(true);
-    }//GEN-LAST:event_jMIInscripcionesActionPerformed
-
-
-    }//GEN-LAST:event_jMIFormularioMateriaActionPerformed
-
-    // CARGO LAS NOTAS DEL ALUMNOS
-    private void jMManipulacionDeNotasActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMManipulacionDeNotasActionPerformed
-        
-        CargaDeNotas ventana=new CargaDeNotas();
-        jPanelPrincipal.removeAll();
-        jPanelPrincipal.repaint();
-        jPanelPrincipal.add(ventana);
-        ventana.setVisible(true);
-        
-    }//GEN-LAST:event_jMManipulacionDeNotasActionPerformed
-
-    private void jMenu47ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMenu47ActionPerformed
-        // TODO add your handling code here:
-       
-
-    }//GEN-LAST:event_jMenu47ActionPerformed
-
-    private void jMenuItem2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMenuItem2ActionPerformed
-        
-        ListadoDeAlumnosPorMateria ventana = new ListadoDeAlumnosPorMateria();
-        jPanelPrincipal.removeAll();
-        jPanelPrincipal.repaint();
-        jPanelPrincipal.add(ventana);
-        ventana.setVisible(true);
-    }//GEN-LAST:event_jMenuItem2ActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(Menu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new Menu().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JMenuItem jMIFormularioAlumno;
-    private javax.swing.JMenuItem jMIFormularioMateria;
-    private javax.swing.JMenuItem jMIInscripciones;
-    private javax.swing.JMenuItem jMManipulacionDeNotas;
-    private javax.swing.JMenu jMenu10;
-    private javax.swing.JMenu jMenu11;
-    private javax.swing.JMenu jMenu12;
-    private javax.swing.JMenu jMenu13;
-    private javax.swing.JMenu jMenu14;
-    private javax.swing.JMenu jMenu15;
-    private javax.swing.JMenu jMenu16;
-    private javax.swing.JMenu jMenu17;
-    private javax.swing.JMenu jMenu18;
-    private javax.swing.JMenu jMenu19;
-    private javax.swing.JMenu jMenu20;
-    private javax.swing.JMenu jMenu21;
-    private javax.swing.JMenu jMenu22;
-    private javax.swing.JMenu jMenu23;
-    private javax.swing.JMenu jMenu24;
-    private javax.swing.JMenu jMenu25;
-    private javax.swing.JMenu jMenu26;
-    private javax.swing.JMenu jMenu27;
-    private javax.swing.JMenu jMenu28;
-    private javax.swing.JMenu jMenu29;
-    private javax.swing.JMenu jMenu3;
-    private javax.swing.JMenu jMenu30;
-    private javax.swing.JMenu jMenu31;
-    private javax.swing.JMenu jMenu32;
-    private javax.swing.JMenu jMenu33;
-    private javax.swing.JMenu jMenu34;
-    private javax.swing.JMenu jMenu35;
-    private javax.swing.JMenu jMenu36;
-    private javax.swing.JMenu jMenu37;
-    private javax.swing.JMenu jMenu38;
-    private javax.swing.JMenu jMenu39;
-    private javax.swing.JMenu jMenu4;
-    private javax.swing.JMenu jMenu40;
-    private javax.swing.JMenu jMenu41;
-    private javax.swing.JMenu jMenu42;
-    private javax.swing.JMenu jMenu43;
-    private javax.swing.JMenu jMenu44;
-    private javax.swing.JMenu jMenu45;
-    private javax.swing.JMenu jMenu46;
-    private javax.swing.JMenu jMenu47;
-    private javax.swing.JMenu jMenu5;
-    private javax.swing.JMenu jMenu6;
-    private javax.swing.JMenu jMenu7;
-    private javax.swing.JMenu jMenu8;
-    private javax.swing.JMenu jMenu9;
-    private javax.swing.JMenuBar jMenuBar1;
-    private javax.swing.JMenuBar jMenuBar2;
-    private javax.swing.JMenuBar jMenuBar3;
-    private javax.swing.JMenuItem jMenuItem1;
-    private javax.swing.JMenuItem jMenuItem2;
-    private javax.swing.JDesktopPane jPanelPrincipal;
-    private javax.swing.JMenu jmAlumnos;
-    private javax.swing.JMenu jmMateria;
-    // End of variables declaration//GEN-END:variables
-}
